Project emf.full.serialization created
In this repository, there will be the projects needed to serialize and
deserialize objects of EMF. Differents methods of serialization will be
located at different projects.

It is needed the miso.emf.serialization project included in the pom
file, which defines the basic interfaces to implement.

Created a new project called emf.full.serialization which will includes
a first attemp of implementation of a EMFSerializer and EMFDeserializer.

More details of the functionality of the serialization will appear
later.